Implemented ToxicityClassifer with BinaryLogistic classifier.

+++ ClassifyBot.Classifier.StanfordNLP/StanfordNLPClassifier.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Diagnostics.Contracts;
 using System.IO;
+using System.Linq;
 using System.Text;
 using System.Text.RegularExpressions;
 
@@ -17,7 +18,14 @@ namespace ClassifyBot
         {
             foreach (KeyValuePair<string, object> kv in AdditionalOptions)
             {
-                ClassifierProperties.Add(kv.Key, kv.Value);
+                if (ClassifierProperties.ContainsKey(kv.Key))
+                {
+                    ClassifierProperties[kv.Key] = kv.Value;
+                }
+                else
+                {
+                    ClassifierProperties.Add(kv.Key, kv.Value);
+                }
                 Info("Using additional classifier property {0}={1}.", kv.Key, kv.Value);
             }
 
@@ -40,10 +48,13 @@ namespace ClassifyBot
             }
  
             ClassifierOutput = new List<string>();
+
             foreach (string s in javaCommand.GetOutputAndErrorLines())
             {
-                if (s.StartsWith("Built this classifier"))
+                if (!BuiltClassifier && s.StartsWith("Built this classifier"))
                 {
+                    BuiltClassifier = true;
+                    
                     Match m = builtClassifierRegex.Match(s);
                     if (m.Success)
                     {
@@ -53,14 +64,24 @@ namespace ClassifyBot
                         NumberofParameters = Int32.Parse(m.Groups[4].Value);
                         Info("Built classifier {0} with {1} features, {2} classes and {3} parameters.", ClassifierType, NumberofFeatures, NumberofClasses, NumberofParameters);
                     }
+                }
+                else if (ClassifierType.Empty() && s.StartsWith("QNMinimizer called on double function"))
+                {
+                    ClassifierType = "BinaryLogisticClassifier";
+                    Match m = binaryClassiferQNN.Match(s);
+                    if (m.Success)
+                    {
+                        NumberofFeatures = Int32.Parse(m.Groups[1].Value);
+                        Info("Built classifier {0} with {1} features.", ClassifierType, NumberofFeatures);
+                    }
                     else
                     {
-                        Error("Could not parse classifier output line: {0}.", s);
-                        return StageResult.FAILED;
+                        Error("Could not parse BinaryLogisticClassifier output: {0}.", s);
                     }
                 }
 
-                else if (s.StartsWith("Reading dataset from {0} ... done".F(TrainingFile.FullName)))
+            
+                else if (!ReadTrainingDataset && s.StartsWith("Reading dataset from {0} ... done".F(TrainingFile.FullName)))
                 {
                     ReadTrainingDataset = true;
                     Match m = readDataSetRegex.Match(s);
@@ -76,7 +97,7 @@ namespace ClassifyBot
                     }
                 }
 
-                else if (s.StartsWith("Reading dataset from {0} ... done".F(TestFile.FullName)))
+                else if (!ReadTestDataset && s.StartsWith("Reading dataset from {0} ... done".F(TestFile.FullName)))
                 {
                     ReadTestDataset = true;
                     Match m = readDataSetRegex.Match(s);
@@ -92,6 +113,45 @@ namespace ClassifyBot
                     }
                 }
 
+                else if (!MicroAveragedF1.HasValue && s.StartsWith("Accuracy/micro-averaged F1"))
+                {
+                    Match m = f1MicroRegex.Match(s);
+                    if (m.Success)
+                    {
+                        MicroAveragedF1 = Single.Parse(m.Groups[1].Value);
+                        Info("Micro-averaged F1 = {0}.", MicroAveragedF1);
+                    }
+                    else
+                    {
+                        Error("Could not parse micro-averaged F1 statistic {0}.", s);
+                    }
+                }
+
+                else if (!MacroAveragedF1.HasValue && s.StartsWith("Macro-averaged F1"))
+                {
+                    Match m = f1MacroRegex.Match(s);
+                    if (m.Success)
+                    {
+                        MacroAveragedF1 = Single.Parse(m.Groups[1].Value);
+                        Info("Macro-averaged F1 = {0}.", MacroAveragedF1);
+                    }
+                    else
+                    {
+                        Error("Could not parse macro-averaged F1 statistic {0}.", s);
+                    }
+                }
+                else if (Features == null && s.StartsWith("Built this classifier: 1"))
+                {
+                    Features = new Dictionary<string, float>();
+                    string f = s.Remove(0,"Built this classifier: ".Length);
+                    foreach (string l in f.Split("/".ToCharArray(), StringSplitOptions.RemoveEmptyEntries))
+                    {
+                        string[] ls = l.Split('=');
+                        Features.Add(ls[0].Trim(), Single.Parse(ls[1].Trim()));
+                    }
+                    Info("Using {0} features.", Features.Count);
+          
+                }
                 else if (s.StartsWith("Cls"))
                 {
                     Match m = classStatisticRegex.Match(s);
@@ -193,6 +253,11 @@ namespace ClassifyBot
                 return StageResult.INVALID_OPTIONS;
             }
 
+            if (BinaryLogisticClassifier && !AdditionalOptions.ContainsKey("useBinary"))
+            {
+                AdditionalOptions.Add("useBinary", true);
+            }
+
             Command version = new Command(Path.Combine(JavaHome, "bin"), "java", "-version");
             Task c = version.Run();
             if (!version.Started)
@@ -219,7 +284,7 @@ namespace ClassifyBot
 
         protected override StageResult Cleanup()
         {
-            if (ClassifierPropsFile != null && ClassifierPropsFile.Exists)
+            if (!KeepPropsFile && ClassifierPropsFile != null && ClassifierPropsFile.Exists)
             {
                 ClassifierPropsFile.Delete();
             }
@@ -232,12 +297,6 @@ namespace ClassifyBot
         #endregion
 
         #region Properties
-        [Option('J', "java-home", Required = false, HelpText = "The path to an existing Java installation. If this is not specified then the JAVA_HOME environment variable will be used")]
-        public virtual string JavaHome { get; set; }
-
-        [Option('C', "class-path", Required = false, HelpText = "The path to the Stanford NLP Classifier jar file. If this is not specified then the JAVA_HOME environment variable will be used")]
-        public virtual string ClassPath { get; set; }
-
         public virtual Dictionary<string, object> ClassifierProperties { get; } = new Dictionary<string, object>()
         {
             {"1.useSplitWords", true },
@@ -248,20 +307,44 @@ namespace ClassifyBot
         };
 
         public FileInfo ClassifierPropsFile { get; protected set; }
+
         public List<string> ClassifierOutput { get; protected set; }
         
+        public int? TrainingDataSetItems { get; protected set; }
 
-        public int TrainingDataSetItems { get; protected set; }
-        public int TestDataSetItems { get; protected set; }
+        public int? TestDataSetItems { get; protected set; }
+
+        public bool BuiltClassifier { get; protected set; }
+
         public bool ReadTrainingDataset { get; protected set; }
-        public bool ReadTestDataset { get; protected set; }
-        public string ClassifierType { get; protected set; }
-        public int NumberofFeatures { get; protected set; }
-        public int NumberofClasses { get; protected set; }
-        public int NumberofParameters { get; protected set; }
 
-        public float MicroAveragedF1 { get; protected set; }
-        public float MacroAveragedF1 { get; protected set; }
+        public bool ReadTestDataset { get; protected set; }
+
+        public string ClassifierType { get; protected set; }
+
+        public int? NumberofFeatures { get; protected set; }
+
+        public int? NumberofClasses { get; protected set; }
+
+        public Dictionary <string, float> Features { get; protected set; }
+
+        public int? NumberofParameters { get; protected set; }
+
+        public float? MicroAveragedF1 { get; protected set; }
+
+        public float? MacroAveragedF1 { get; protected set; }
+
+        [Option('j', "java-home", Required = false, HelpText = "The path to an existing Java installation. If this is not specified then the JAVA_HOME environment variable will be used")]
+        public virtual string JavaHome { get; set; }
+
+        [Option('c', "class-path", Required = false, HelpText = "The path to the Stanford NLP Classifier jar file. If this is not specified then the JAVA_HOME environment variable will be used")]
+        public virtual string ClassPath { get; set; }
+
+        [Option('k', "keep-props-file", Required = false, Default = false, HelpText = "Don't delete the classifier properties file after classification task completes.")]
+        public bool KeepPropsFile { get; set; }
+
+        [Option('b', "binary", Required = false, Default = false, HelpText = "Use a binary logistic classifier.")]
+        public bool BinaryLogisticClassifier { get; set; }
         #endregion
 
         #region Methods
@@ -281,10 +364,13 @@ namespace ClassifyBot
 
         #region Fields
         protected JavaCommand javaCommand;
-        private static Regex builtClassifierRegex = new Regex("Built this classifier: (\\S+) with (\\d+) features, (\\d+) classes, and (\\d+) parameters.", RegexOptions.Compiled);
-        private static Regex readDataSetRegex = new Regex("Reading dataset from (.+)done \\[(\\d+\\.\\d+)s, (\\d+) items\\]", RegexOptions.Compiled);
-        private static Regex classStatisticRegex = new Regex("Cls (\\S+): TP=(\\d+) FN=(\\d+) FP=(\\d+) TN=(\\d+); Acc (\\d+.\\d+) P (\\d+.\\d+) R (\\d+.\\d+) F1 (\\d+.\\d+)", RegexOptions.Compiled);
-        private static Regex resultRegex = new Regex("(\\S+)\\s+(\\S+)\\s+(\\d+.\\d+)\\s+(\\d+.\\d+)", RegexOptions.Compiled);
+        protected static Regex builtClassifierRegex = new Regex("Built this classifier: (\\S+) with (\\d+) features, (\\d+) classes, and (\\d+) parameters.", RegexOptions.Compiled);
+        protected static Regex readDataSetRegex = new Regex("Reading dataset from (.+)done \\[(\\d+\\.\\d+)s, (\\d+) items\\]", RegexOptions.Compiled);
+        protected static Regex classStatisticRegex = new Regex("Cls (\\S+): TP=(\\d+) FN=(\\d+) FP=(\\d+) TN=(\\d+); Acc (\\d+.\\d+) P (\\d+.\\d+) R (\\d+.\\d+) F1 (\\d+.\\d+)", RegexOptions.Compiled);
+        protected static Regex resultRegex = new Regex("(\\S+)\\s+(\\S+)\\s+(\\d+.\\d+)\\s+(\\d+.\\d+)", RegexOptions.Compiled);
+        protected static Regex f1MicroRegex = new Regex("Accuracy/micro-averaged F1: ([0|1]\\.\\d+)");
+        protected static Regex f1MacroRegex = new Regex("Macro-averaged F1: ([0|1]\\.\\d+)");
+        protected static Regex binaryClassiferQNN = new Regex("QNMinimizer called on double function of (\\d+) variables");
         #endregion
     }
 }
